+++ src/openMVG/matching_image_collection/F_ACRobust.hpp
@@ -22,7 +22,7 @@ namespace openMVG {
 
 namespace sfm {
   struct Regions_Provider;
-} // namespace sfm 
+} // namespace sfm
 
 namespace matching_image_collection {
 
@@ -78,7 +78,7 @@ struct GeometricFilter_FMatrix_AC
 
     // Robustly estimate the Fundamental matrix with A Contrario ransac
     const double upper_bound_precision = Square(m_dPrecision);
-    std::vector<size_t> vec_inliers;
+    std::vector<uint32_t> vec_inliers;
     const std::pair<double,double> ACRansacOut =
       ACRANSAC(kernel, vec_inliers, m_stIteration, &m_F, upper_bound_precision);
 
@@ -86,7 +86,7 @@ struct GeometricFilter_FMatrix_AC
       m_dPrecision_robust = ACRansacOut.first;
       // update geometric_inliers
       geometric_inliers.reserve(vec_inliers.size());
-      for ( const size_t & index : vec_inliers)  {
+      for (const uint32_t & index : vec_inliers) {
         geometric_inliers.push_back( vec_PutativeMatches[index] );
       }
       return true;
