+++ src/openMVG_Samples/multiview_robust_estimation_tutorial/robust_estimation_tutorial_homography.cpp
@@ -57,7 +57,7 @@ void display_info
   const PointFeatures & featsR,
   const IndMatches & vec_PutativeMatches,
   const Mat3 &H,
-  const std::vector<size_t> & vec_inliers,
+  const std::vector<uint32_t> & vec_inliers,
   const std::string & sMethod
 );
 
@@ -174,7 +174,7 @@ int main() {
     // The Model type
     Mat3 H;
     // The inlier list
-    std::vector<size_t> vec_inliers;
+    std::vector<uint32_t> vec_inliers;
     H =
       MaxConsensus
       (
@@ -207,7 +207,7 @@ int main() {
     // The Model type
     Mat3 H;
     // The inlier list
-    std::vector<size_t> vec_inliers;
+    std::vector<uint32_t> vec_inliers;
     // Inlier count
     size_t inlier_count = 0;
     H =
@@ -249,7 +249,7 @@ int main() {
         &lmeds_threshold // Upper bound of the tolerated error for the inliers
       );
     // List the inliers
-    std::vector<size_t> vec_inliers;
+    std::vector<uint32_t> vec_inliers;
     for (int i = 0; i < xL.cols(); ++i)
     {
       const double residual_error = std::sqrt(kernel.Error(i, H));
@@ -289,7 +289,7 @@ int main() {
     // The Model type
     Mat3 H;
     // The inlier list
-    std::vector<size_t> vec_inliers;
+    std::vector<uint32_t> vec_inliers;
     // Call the Robust Estimator on the KERNEL
     const std::pair<double,double> ACRansacOut = // Return the precision & the associated NFA
       ACRANSAC(
@@ -322,7 +322,7 @@ void display_info
   const PointFeatures & featsR,
   const IndMatches & vec_PutativeMatches,
   const Mat3 &H,
-  const std::vector<size_t> & vec_inliers,
+  const std::vector<uint32_t> & vec_inliers,
   const std::string & sMethod
 )
 {
