+++ src/openMVG_Samples/multiview_robust_fundamental/robust_fundamental.cpp
@@ -117,7 +117,7 @@ int main() {
     }
 
     //-- Fundamental robust estimation
-    std::vector<size_t> vec_inliers;
+    std::vector<uint32_t> vec_inliers;
     using KernelType =
       ACKernelAdaptor<
         openMVG::fundamental::kernel::SevenPointSolver,
